pyproject.toml
[tool.poetry]
name = "asterisk-prometheus-exporter"
version = "1.0.0"
description = "Prometheus exporter for Asterisk using the Asterisk Manager Interface (AMI)"
authors = ["GONICUS GmbH"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "=3.8.10"
asterisk-ami = "=0.1.6"
prometheus-client = "=0.16.0"
pyyaml = "^6.0"
jsonschema = "^4.17.3"

[build-system]
requires = ["poetry-core=1.3.2"]
build-backend = "poetry.core.masonry.api"
